Removed node1 and node2. Renamed test as imu_sensor. Added publishing code to imu_sensor. Modified UDPMessage data size since the size was too small for containing Imu message.
