+++ src/apps/nodes.h
@@ -1,10 +1,8 @@
 #ifndef ASW_APPS_APPLICATION_TASKS_H_
 #define ASW_APPS_APPLICATION_TASKS_H_
 
-#include "nodes/node1/node1.h"
-#include "nodes/node2/node2.h"
 #include "nodes/ultrasonic_sensor/ultrasonic_sensor.h"
-#include "nodes/test/test.h"
+#include "nodes/imu_sensor/imu_sensor.h"
 
 typedef struct node_descriptor {
 	char name[32];
@@ -14,9 +12,7 @@ typedef struct node_descriptor {
 
 node_decriptor nodes[] = {
 		{"ultrasonic_sensor", ultrasonic_sensor},
-		{"test", test},
-		//{"node1", node1},
-		//{"node2", node2}
+		{"imu_sensor", imu_sensor},
 };
