+++ src/apps/nodes/imu_sensor/imu_sensor.h
@@ -1,0 +1,1 @@
+void imu_sensor(void* p);
