--- src/apps/nodes/node1/node1.h
@@ -1,1 +1,0 @@
-void node1(void* params);
