--- src/apps/nodes/node2/node2.h
@@ -1,1 +1,0 @@
-void node2(void* params);
