--- src/apps/nodes/test/test.h
@@ -1,1 +1,0 @@
-void test(void* p);
