+++ src/rmw/TopicWriter.cpp
@@ -20,7 +20,7 @@ TopicWriter::TopicWriter(const char* callerID, const char* topic, const char* ms
 }
 void TopicWriter::serializeMsg(const ros::Msg& msg, unsigned char* outbuffer)
 {
-	unsigned char stream1[100];
+	unsigned char stream1[sizeof(UDPMessage)];
 	uint32_t offset = msg.serialize(stream1);
 	memcpy(outbuffer, &offset, sizeof(uint32_t));
 	memcpy(outbuffer+sizeof(uint32_t), stream1, offset);
